requirements.txt
opik
openai
weaviate-client
python-dotenv
beautifulsoup4
requests
